requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
